+++ docs/changelog.rst
@@ -27,7 +27,7 @@ Also:
 
 * Reduced verbosity of "Failing the WebSocket connection" logs.
 
-* Allowed `extra_headers` parameter to override User-Agent header.
+* Allowed ``extra_headers`` to override ``Server`` and ``User-Agent`` headers.
 
 3.4
 ...